test(PayloadManager): Moved 2 tests relating to PaylaodManager to same class. Added todo.
